Fixtures: make the `entry_points` fixture publicly available
This fixture makes it easy to temporarily add or remove entry points
which comes in very handy when testing custom implementations of various
classes that are pluginnable. By moving them from `tests/conftest.py` to
the `aiida.manage.tests.pytest_fixtures` module, it can also be used by
plugin packages.
